requirements.txt
streamlit
opencv-python-headless
pillow
bcrypt
razorpay
pandas
numpy
